+++ src/app/components/NotionRenderer/components/Code.tsx
@@ -1,7 +1,4 @@
-"use client";
-
 import { CodeBlock } from "../types";
-import { useState } from "react";
 import prism from "prismjs";
 import "prismjs/components/prism-clike.min.js";
 import "prismjs/components/prism-css-extras.min.js";
@@ -12,6 +9,7 @@ import "prismjs/components/prism-json.min.js";
 import "prismjs/components/prism-jsx.min.js";
 import "prismjs/components/prism-tsx.min.js";
 import "prismjs/components/prism-typescript.min.js";
+import { CopyButton } from "@/app/components/NotionRenderer/components/CopyButton";
 
 interface CodeProps {
   block: CodeBlock;
@@ -21,12 +19,11 @@ export function Code({ block }: CodeProps) {
   const codeData = block.code;
   const richText = codeData.rich_text;
   const language = codeData.language || "typescript";
-  const [isCopied, setIsCopied] = useState(false);
 
   // Get the plain text content
   const codeContent = richText.map((rt) => rt.plain_text).join("");
 
-  // Apply prism syntax highlighting statically
+  // Apply prism syntax highlighting statically during server-side render
   const highlightedCode = (() => {
     try {
       return prism.highlight(
@@ -40,26 +37,9 @@ export function Code({ block }: CodeProps) {
     }
   })();
 
-  const handleCopy = () => {
-    navigator.clipboard
-      .writeText(codeContent)
-      .then(() => {
-        setIsCopied(true);
-        setTimeout(() => setIsCopied(false), 2000);
-      })
-      .catch((err) => {
-        console.error("Failed to copy:", err);
-      });
-  };
-
   return (
     <div className="notion-code">
-      <button
-        className="notion-code-copy notion-code-copy-button"
-        onClick={handleCopy}
-      >
-        {isCopied ? "Copied!" : "📋"}
-      </button>
+      <CopyButton codeContent={codeContent} />
       <pre className={`language-${language}`}>
         <code
           className={`language-${language}`}
